Move storage engine abstract class to types file, add ServerStorageEngine
