+++ src/Login.tsx
@@ -11,7 +11,7 @@ import { PREFIX } from './utils/Prefix';
 import { useAuth } from './store/hooks/useAuth';
 import { useStorageEngine } from './storage/storageEngineHooks';
 import { FirebaseStorageEngine } from './storage/engines/FirebaseStorageEngine';
-import { StorageEngine } from './storage/engines/StorageEngine';
+import { StorageEngine } from './storage/engines/types';
 import { showNotification } from './utils/notifications';
 
 export async function signInWithGoogle(storageEngine: StorageEngine | undefined, setLoading: (val: boolean) => void) {
