+++ src/components/interface/AnalysisFooter.tsx
@@ -10,7 +10,7 @@ import {
 import { useAsync } from '../../store/hooks/useAsync';
 
 import { useStorageEngine } from '../../storage/storageEngineHooks';
-import { StorageEngine } from '../../storage/engines/StorageEngine';
+import { StorageEngine } from '../../storage/engines/types';
 import { useCurrentComponent, useCurrentStep } from '../../routes/utils';
 import { encryptIndex } from '../../utils/encryptDecryptIndex';
 import {
