+++ src/utils/nextParticipant.ts
@@ -1,5 +1,5 @@
 import { StudyConfig } from '../parser/types';
-import { StorageEngine } from '../storage/engines/StorageEngine';
+import { StorageEngine } from '../storage/engines/types';
 import { ParticipantMetadata } from '../store/types';
 
 export function getNewParticipant(storageEngine: StorageEngine | undefined, studyConfig: StudyConfig, metadata: ParticipantMetadata, studyHref: string) {
